Split out postcard.jsx
