--- frontend/journal-app/src/components/elements/PostCard.module.css
@@ -1,214 +1,0 @@
-.postCard {
-    width: 60%;
-    min-height: 8rem;
-    margin-bottom: 1rem;
-
-    border-radius: 0.5rem;
-}
-
-.header {
-    width: 100%;
-    height: 4rem;
-
-    background-color: var(--primary-color);
-    border-radius: 0.5rem 0.5rem 0 0;
-
-    display: flex;
-}
-
-.headerText {
-    width: 50%;
-    padding: 0.4rem 0.5rem;
-
-    display: flex;
-    flex-direction: column;
-    justify-content: start;
-}
-
-.headerText h3 {
-    color: var(--tertiary-color);
-    font-weight: 300;
-}
-
-.headerText p {
-    color: var(--tertiary-color);
-    font-size: 0.7rem;
-}
-
-.headerTools {
-    width: 50%;
-    padding: 0.1rem 0.5rem;
-
-    display: flex;
-    justify-content: end;
-}
-
-.tool {
-    height: 1rem;
-    margin-left: 0.5rem;
-
-    color: var(--tertiary-color);
-    cursor: pointer;
-}
-
-.tool:nth-child(2) {
-    font-size: 1.3rem;
-
-    position: relative;
-    bottom: 0.3rem;
-}
-
-.content {
-    height: calc(100% - 4rem);
-
-    border-left: 1px solid var(--primary-color);
-    border-right: 1px solid var(--primary-color);
-    border-bottom: 1px solid var(--primary-color);
-    border-radius: 0 0 0.5rem 0.5rem;
-
-    display: flex;
-    flex-direction: column;
-}
-
-.textarea {
-    width: 100%;
-    height: 100%;
-    padding: 0.5rem;
-
-    outline: none;
-    border: none;
-    border-radius: 0.5rem;
-
-    resize: none;
-}
-
-.textarea::-webkit-scrollbar,
-.newTextarea::-webkit-scrollbar {
-    display: none;
-}
-
-.modalOverlay {
-    position: fixed;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    background-color: rgba(0, 0, 0, 0.5);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    z-index: 1000;
-}
-
-.openModalButton {
-    padding: 0.25rem 1.5rem;
-
-    color: var(--tertiary-color);
-    background-color: var(--primary-color);
-    font-family: "poppins", "sans-serif";
-    outline: none;
-    border: none;
-    border-radius: 0.25rem;
-    cursor: pointer;
-}
-
-.newCard {
-    width: 40%;
-    height: 25rem;
-    padding: 0rem 1rem 1rem 1rem;
-
-    background-color: var(--primary-color);
-    border-radius: 0.5rem;
-
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-}
-
-.newHeader {
-    width: 100%;
-    height: 15%;
-
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-
-.newHeaderTools {
-    width: 30%;
-
-    display: flex;
-    justify-content: end;
-}
-
-.newTextarea {
-    width: 100%;
-    height: 85%;
-    padding: 0.5rem;
-
-    outline: none;
-    border: none;
-    border-radius: 0.5rem;
-
-    resize: none;
-}
-
-.deleteCard {
-    width: 15rem;
-    height: 10rem;
-
-    background-color: white;
-    border-radius: 0.5rem;
-
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-}
-
-.deleteCard div {
-    margin-top: 1rem;
-    width: 8.5rem;
-
-    display: flex;
-    justify-content: space-between;
-}
-
-button {
-    background-color: var(--secondary-color);
-    color: var(--tertiary-color);
-    font-family: "poppins", "sans-serif";
-    border-radius: 0.25rem;
-    border: 1px solid var(--tertiary-color);
-    outline: none;
-    cursor: pointer;
-}
-
-.deleteCard button {
-    height: 2rem;
-    width: 4rem;
-}
-
-.newHeaderTools button {
-    height: 1.5rem;
-    margin-left: 0.5rem;
-    padding: 0rem 1rem;
-}
-
-.titleInput{
-    height: 2rem;
-    width: 70%;
-
-    background-color: rgb(0, 0, 0, 0);
-    color: var(--tertiary-color);
-    font-family: "poppins", "sans-serif";
-    font-size: 1.16rem;
-    font-weight: 300;
-    outline: none;
-    border: none;
-    border-bottom: 1px solid var(--tertiary-color);
-}
-
-.titleInput::placeholder {
-    color: rgb(170, 170, 170);
-}
