+++ frontend/journal-app/src/components/postComponents/DeletePostModal.module.css
@@ -1,0 +1,56 @@
+.modalOverlay {
+    position: fixed;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    background-color: rgba(0, 0, 0, 0.5);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    z-index: 1000;
+}
+
+.deleteCard {
+    width: 15rem;
+    height: 10rem;
+
+    background-color: white;
+    border-radius: 0.5rem;
+
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+}
+
+.deleteCard div {
+    margin-top: 1rem;
+    width: 8.5rem;
+
+    display: flex;
+    justify-content: space-between;
+}
+
+button {
+    background-color: var(--secondary-color);
+    color: var(--tertiary-color);
+    font-family: "poppins", "sans-serif";
+    border-radius: 0.25rem;
+    border: 1px solid var(--tertiary-color);
+    outline: none;
+    cursor: pointer;
+}
+
+.deleteCard button {
+    height: 2rem;
+    width: 4rem;
+}
+
+.tool {
+    height: 1rem;
+    margin-left: 0.5rem;
+
+    color: var(--tertiary-color);
+    cursor: pointer;
+}
