+++ frontend/journal-app/src/components/postComponents/PostCard.module.css
@@ -1,0 +1,88 @@
+.postCard {
+    width: 60%;
+    min-height: 8rem;
+    margin-bottom: 1rem;
+
+    border-radius: 0.5rem;
+}
+
+.header {
+    width: 100%;
+    height: 4rem;
+
+    background-color: var(--primary-color);
+    border-radius: 0.5rem 0.5rem 0 0;
+
+    display: flex;
+}
+
+.headerText {
+    width: 50%;
+    padding: 0.4rem 0.5rem;
+
+    display: flex;
+    flex-direction: column;
+    justify-content: start;
+}
+
+.headerText h3 {
+    color: var(--tertiary-color);
+    font-weight: 300;
+}
+
+.headerText p {
+    color: var(--tertiary-color);
+    font-size: 0.7rem;
+}
+
+.headerTools {
+    width: 50%;
+    padding: 0.1rem 0.5rem;
+
+    display: flex;
+    justify-content: end;
+}
+
+.tool {
+    height: 1rem;
+    margin-left: 0.5rem;
+
+    color: var(--tertiary-color);
+    cursor: pointer;
+}
+
+.tool:nth-child(2) {
+    font-size: 1.3rem;
+
+    position: relative;
+    bottom: 0.3rem;
+}
+
+.content {
+    height: calc(100% - 4rem);
+
+    border-left: 1px solid var(--primary-color);
+    border-right: 1px solid var(--primary-color);
+    border-bottom: 1px solid var(--primary-color);
+    border-radius: 0 0 0.5rem 0.5rem;
+
+    display: flex;
+    flex-direction: column;
+}
+
+.textarea {
+    width: 100%;
+    height: 100%;
+    padding: 0.5rem;
+
+    outline: none;
+    border: none;
+    border-radius: 0.5rem;
+
+    resize: none;
+}
+
+.textarea::-webkit-scrollbar,
+.newTextarea::-webkit-scrollbar {
+    display: none;
+}
